--- a/pyproject.toml
+++ b/pyproject.toml
@@ -17,7 +17,6 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.11, <3.12"
-pandas = "^2.1.3"
 pydantic-settings = "^2.0.3"
 pymupdf = "^1.23.7"
 pdftotext = "^2.2.2"
@@ -42,6 +41,7 @@ pylint-pydantic = "^0.3.0"
 mypy = "^1.6.1"
 types-tqdm = "^4.66.0.5"
 types-dateparser = "^1.2.0.20240420"
+types-tabulate = "^0.9.0.20240106"
 
 
 [tool.taskipy.tasks]