--- a/pyproject.toml
+++ b/pyproject.toml
@@ -22,7 +22,6 @@ pymupdf = "^1.23.7"
 pdftotext = "^2.2.2"
 click = "^8.1.7"
 tqdm = "^4.66.1"
-pdf2john = "^0.2.0"
 tabulate = "^0.9.0"
 mypy = "^1.7.1"
 pydantic = "^2.5.2"
@@ -85,7 +84,6 @@ module = [
     "fitz",
     "pdftotext",
     "pdf2john",
-    "pandas",
 ]
 ignore_missing_imports = true