--- a/src/monopoly/examples/single_statement.py
+++ b/src/monopoly/examples/single_statement.py
@@ -13,15 +13,17 @@ def example():
     statement = handler.extract()
 
     # Dates are converted into an ISO 8601 date format
-    transformed_df = handler.transform(
-        df=statement.df,
+    transactions = handler.transform(
+        transactions=statement.transactions,
         statement_date=statement.statement_date,
         transaction_date_order=statement.config.transaction_date_order,
     )
 
     # Parsed transactions writen to a CSV file in the "example" directory
     handler.load(
-        df=transformed_df, statement=statement, output_directory="src/monopoly/examples"
+        transactions=transactions,
+        statement=statement,
+        output_directory="src/monopoly/examples",
     )