--- a/src/monopoly/pdf.py
+++ b/src/monopoly/pdf.py
@@ -1,5 +1,5 @@
 import logging
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from functools import cached_property, lru_cache
 from io import BytesIO
 from pathlib import Path
@@ -7,13 +7,31 @@ from typing import Any, Optional
 
 import fitz
 import pdftotext
-from pdf2john import PdfHashExtractor as EncryptionMetadataExtractor
 from pydantic import SecretStr
 
 from monopoly.banks import detect_bank
 from monopoly.constants import EncryptionIdentifier, MetadataIdentifier
 
 logger = logging.getLogger(__name__)
+
+
+@dataclass
+class EncryptDict:
+    """Stores encryption dictionary of a PDF"""
+
+    raw_encrypt_dict: dict
+    pdf_version: str
+    algorithm: int = field(init=False)
+    length: int = field(init=False)
+    permissions: int = field(init=False)
+    revision: int = field(init=False)
+
+    def __post_init__(self):
+        self.pdf_version = float(self.pdf_version[-3:])
+        self.algorithm = int(self.raw_encrypt_dict.get("V"))
+        self.length = int(self.raw_encrypt_dict.get("Length"))
+        self.permissions = int(self.raw_encrypt_dict.get("P"))
+        self.revision = int(self.raw_encrypt_dict.get("R"))
 
 
 @dataclass
@@ -46,7 +64,7 @@ class BadPasswordFormatError(Exception):
 class PdfParser:
     def __init__(
         self,
-        passwords: Optional[list[str]] = None,
+        passwords: Optional[list[SecretStr]] = None,
         file_path: Optional[Path] = None,
         file_bytes: Optional[bytes] = None,
     ):
@@ -89,16 +107,13 @@ class PdfParser:
         """
         identifiers = []
         # pylint: disable=protected-access
-        if self.extractor.encrypt_dict:
-            extractor = self.extractor
-            major, minor = extractor.pdf._header_version
-            pdf_version = f"{major}.{minor}"
+        if encrypt_dict := self.encrypt_dict:
             encryption_identifier = EncryptionIdentifier(
-                float(pdf_version),
-                extractor.algorithm,
-                extractor.revision,
-                extractor.length,
-                extractor.permissions,
+                float(encrypt_dict.pdf_version),
+                encrypt_dict.algorithm,
+                encrypt_dict.revision,
+                encrypt_dict.length,
+                encrypt_dict.permissions,
             )
             identifiers.append(encryption_identifier)
 
@@ -176,27 +191,27 @@ class PdfParser:
         """
         Returns a Python representation of a PDF document.
         """
-        if self.file_path:
-            return fitz.Document(filename=self.file_path)
+        if not self.file_path and not self.file_bytes:
+            raise RuntimeError("Either `file_path` or `file_bytes` must be passed")
 
-        if self.file_bytes:
-            return fitz.Document(stream=self.file_bytes)
+        if self.file_path and self.file_bytes:
+            raise RuntimeError(
+                "Only one of `file_path` or `file_bytes` should be defined"
+            )
 
-        raise RuntimeError("Either file path or file bytestream must be passed")
+        args = {"filename": self.file_path, "stream": self.file_bytes}
+        return fitz.Document(**args)
 
     @cached_property
-    def extractor(self) -> EncryptionMetadataExtractor:
-        """
-        Returns an instance of pdf2john, used to retrieve and return
-        the encryption metadata from a PDF's encryption dictionary
-        """
-        if self.file_path:
-            return EncryptionMetadataExtractor(file_name=self.file_path)
+    def encrypt_dict(self) -> EncryptDict | None:
+        stream = self._get_doc_byte_stream()
+        pdf_version_string = stream.read(8).decode("utf-8", "backslashreplace")
 
-        if self.file_bytes:
-            return EncryptionMetadataExtractor(file_bytes=self.file_bytes)
-
-        raise RuntimeError("Either file path or file bytestream must be passed")
+        if self.document.is_encrypted:
+            raw_encrypt_dict = self._get_raw_encrypt_dict(self.document)
+            encrypt_dict = EncryptDict(raw_encrypt_dict, pdf_version_string)
+            return encrypt_dict
+        return None
 
     @staticmethod
     def _remove_vertical_text(page: fitz.Page):
@@ -224,3 +239,26 @@ class PdfParser:
                     page.add_redact_annot(line["bbox"])
         page.apply_redactions(images=fitz.PDF_REDACT_IMAGE_NONE)
         return page
+
+    @staticmethod
+    def _get_raw_encrypt_dict(doc: fitz.Document) -> dict:
+        """
+        Helper function to extract the PDF encryption dictionary, since
+        `fitz` doesn't provide it
+        """
+        encrypt_metadata = {}
+        pdf_object, value = doc.xref_get_key(-1, "Encrypt")
+        if pdf_object != "xref":
+            pass  # PDF has no metadata
+        else:
+            xref = int(value.replace("0 R", ""))  # extract the metadata xref
+            for key in doc.xref_get_keys(xref):
+                encrypt_metadata[key] = doc.xref_get_key(xref, key)[1]
+        return encrypt_metadata
+
+    def _get_doc_byte_stream(self) -> BytesIO:
+        if self.file_path:
+            with open(self.file_path, "rb") as file:
+                stream = BytesIO(file.read())
+            return stream
+        raise RuntimeError("Unable to create stream since `file_path` not passed")