--- a/tests/integration/banks/test_banks_credit.py
+++ b/tests/integration/banks/test_banks_credit.py
@@ -1,14 +1,12 @@
 from datetime import datetime
 from pathlib import Path
 
-import pandas as pd
 import pytest
-from pandas.testing import assert_frame_equal
 from test_utils.skip import skip_if_encrypted
+from test_utils.transactions import get_transactions_as_dict, read_transactions_from_csv
 
 from monopoly.banks import Citibank, Dbs, Hsbc, Ocbc, StandardChartered
 from monopoly.banks.base import BankBase
-from monopoly.constants import StatementFields
 from monopoly.handler import StatementHandler
 from monopoly.statements import CreditStatement
 
@@ -35,18 +33,26 @@ def test_bank_credit_statements(
     statement: CreditStatement = statement_handler.extract()
 
     # check raw data
-    expected_raw_data = pd.read_csv(test_directory / "raw.csv")
-    df = statement.df
-    assert_frame_equal(df, expected_raw_data)
-    assert round(df[StatementFields.AMOUNT].sum(), 2) == total_amount
+    expected_raw_transactions = read_transactions_from_csv(test_directory, "raw.csv")
+    raw_transactions_as_dict = get_transactions_as_dict(statement.transactions)
+    expected_transaction_total_amount = [
+        transaction.amount for transaction in statement.transactions
+    ]
+    assert expected_raw_transactions == raw_transactions_as_dict
+    assert round(sum(expected_transaction_total_amount), 2) == total_amount
     assert statement.statement_date == statement_date
     assert statement.perform_safety_check()
 
     # check transformed data
-    expected_transformed_data = pd.read_csv(test_directory / "transformed.csv")
-    transformed_df = statement_handler.transform(
-        df=df,
+    expected_transformed_transactions = read_transactions_from_csv(
+        test_directory, "transformed.csv"
+    )
+    transformed_transactions = statement_handler.transform(
+        transactions=statement.transactions,
         statement_date=statement.statement_date,
         transaction_date_order=statement.config.transaction_date_order,
     )
-    assert_frame_equal(transformed_df, expected_transformed_data)
+    transformed_transactions_as_dict = get_transactions_as_dict(
+        transformed_transactions
+    )
+    assert expected_transformed_transactions == transformed_transactions_as_dict