- Moved the statement handler import statement inside the CLI's `process_statement()` function, which greatly reduces the latency when calling the CLI without any operations (~1300ms -> 175ms)
- Removed unnecessary pandas and pdf2john dependencies
```
**Original results:**
Benchmark 1: monopoly src/monopoly/examples/example_statement.pdf -s
  Time (mean ± σ):      1.350 s ±  0.087 s    [User: 1.452 s, System: 0.685 s]
  Range (min … max):    1.245 s …  1.544 s    10 runs
Benchmark 2: monopoly statements
  Time (mean ± σ):      3.230 s ±  0.441 s    [User: 19.590 s, System: 1.492 s]
  Range (min … max):    2.891 s …  4.344 s    10 runs

**New results (without pandas):**
Benchmark 1: monopoly src/monopoly/examples/example_statement.pdf -s
  Time (mean ± σ):      1.134 s ±  0.116 s    [User: 1.074 s, System: 0.058 s]
  Range (min … max):    0.978 s …  1.390 s    10 runs
Benchmark 2: monopoly statements
  Time (mean ± σ):      3.110 s ±  0.282 s    [User: 31.059 s, System: 1.337 s]
  Range (min … max):    2.784 s …  3.712 s    10 runs
  
**New results (without pandas & pdf2john):**
Benchmark 1: monopoly src/monopoly/examples/example_statement.pdf -s
  Time (mean ± σ):     734.3 ms ±  54.6 ms    [User: 703.4 ms, System: 25.9 ms]
  Range (min … max):   688.3 ms … 859.5 ms    10 runs
Benchmark 2: monopoly statements
  Time (mean ± σ):      2.754 s ±  0.260 s    [User: 26.748 s, System: 1.012 s]
  Range (min … max):    2.462 s …  3.194 s    10 runs
```